+++ src/core/utilities/envuments/Envuments.ts
@@ -1,22 +1,49 @@
-import * as dotenv from 'dotenv';
-
-// eslint-disable-next-line import/no-cycle
-import { Parser } from './lib/Parser';
-import { EnvumentType } from './lib/Types.enum';
+import { config } from 'dotenv';
 
 let configObject: Record<string, unknown> = {};
 
-export class Envuments {
+class Parser {
+  private readonly TEMPLATE_REGEX = /\${\w*}/g;
+
+  private escapeRegexChars(str: string): string {
+    return str.replace(/([.*+?^=!:${}()|[\]/\\])/g, '\\$1');
+  }
+
+  resolveValueString(key: string, value: string): string {
+    const templates = value.match(this.TEMPLATE_REGEX);
+    if (!templates) return value;
+
+    for (const template of templates) {
+      const variable = template.slice(2, -1);
+      if (!variable || variable === key) continue; // Prevent any circulars
+
+      const variableVal = Envuments.get(variable);
+
+      value = value.replace(new RegExp(this.escapeRegexChars(template), 'g'), variableVal || template);
+    }
+
+    return value;
+  }
+}
+
+enum EnvumentType {
+  String,
+  Number,
+  Boolean
+}
+
+class Envuments {
   private static readonly parser = new Parser();
 
   private static getConfig(): Record<string, unknown> {
     if (!Object.keys(configObject).length) {
       // Default to dotenv config
       try {
-        dotenv.config();
+        config({ quiet: true });
       } catch {
         // Do Nothing
       }
+
       Envuments.seedConfig(process.env);
     }
 
@@ -32,22 +59,22 @@ export class Envuments {
     }
   }
 
-  private static _get(key: string, type: EnvumentType = EnvumentType.STRING, def?: unknown): unknown {
+  private static _get(key: string, type: EnvumentType = EnvumentType.String, def?: unknown): unknown {
     const rawVal = this.getConfig()[key] as string | number | boolean;
     if (!rawVal) return def;
 
     const parsed = this.parser.resolveValueString(key, String(rawVal));
 
     switch (type) {
-      case EnvumentType.STRING: {
+      case EnvumentType.String: {
         return (parsed !== '' && parsed) || def;
       }
-      case EnvumentType.NUMBER: {
+      case EnvumentType.Number: {
         const num = Number(parsed);
 
         return (!isNaN(num) && num) || def;
       }
-      case EnvumentType.BOOLEAN: {
+      case EnvumentType.Boolean: {
         const yes = ['1', 'yes', 'true'];
         const no = ['0', 'no', 'false'];
 
@@ -64,26 +91,28 @@ export class Envuments {
   }
 
   static get(key: string, def?: string): string {
-    return this._get(key, EnvumentType.STRING, def) as string;
+    return this._get(key, EnvumentType.String, def) as string;
   }
 
   static getNumber(key: string, def?: number): number {
-    return this._get(key, EnvumentType.NUMBER, def) as number;
+    return this._get(key, EnvumentType.Number, def) as number;
   }
 
   static getBoolean(key: string, def?: boolean): boolean {
-    return this._get(key, EnvumentType.BOOLEAN, def) as boolean;
+    return this._get(key, EnvumentType.Boolean, def) as boolean;
   }
 
   get(key: string, def?: string): string {
-    return Envuments._get(key, EnvumentType.STRING, def) as string;
+    return Envuments._get(key, EnvumentType.String, def) as string;
   }
 
   getNumber(key: string, def?: number): number {
-    return Envuments._get(key, EnvumentType.NUMBER, def) as number;
+    return Envuments._get(key, EnvumentType.Number, def) as number;
   }
 
   getBoolean(key: string, def?: boolean): boolean {
-    return Envuments._get(key, EnvumentType.BOOLEAN, def) as boolean;
+    return Envuments._get(key, EnvumentType.Boolean, def) as boolean;
   }
 }
+
+export { Envuments, EnvumentType, Parser };
