--- src/core/utilities/envuments/lib/Types.enum.ts
@@ -1,5 +1,0 @@
-export enum EnvumentType {
-  STRING,
-  NUMBER,
-  BOOLEAN
-}
